:recycle: Move the offer functions to file offer-utilities.ts and move corresponding tests
